Rename api to routes in admin-service
